requirements.txt
google-cloud-translate
google-cloud-language
google-cloud-bigquery
google-cloud-logging
tweepy
psaw
